Refactor to move placement event handlers
